se modifico para que ahora almacene en json

+++ buscar_datos.py
@@ -1,20 +1,50 @@
 from bs4 import BeautifulSoup
 import requests
 import os
+import json
+from producto import Producto
 # Este script busca productos en Mercado Libre Argentina y guarda los datos en un archivo de texto.
 
 #URL_BASE = "https://listado.mercadolibre.com.ar/harry-potter#D[A:harry%20potter]"
 
-def buscar_datos(nombre_archivo="datos.txt"):
-    a_buscar = input("Ingrese a buscar: ").replace(" ", "-")
-    URL_BASE = f"https://listado.mercadolibre.com.ar/{a_buscar}"
-    datos_obtenido = requests.get(URL_BASE)
+datos_empresa = {
+  "nombre": "TechNova Components",
+  "horarios_atencion": {
+    "lunes_viernes": "09:00 - 18:00",
+    "sabado": "10:00 - 14:00",
+    "domingo": "Cerrado"
+  },
+  "telefono": "+54 11 3456-7890",
+  "correo": "contacto@technova.com",
+  "productos": [],
+  "empleados": [
+    {
+      "nombre": "Lucía Fernandez",
+      "telefono": "+54 11 6543-2100"
+    },
+    {
+      "nombre": "Carlos Gomez",
+      "telefono": "+54 11 6789-1234"
+    },
+    {
+      "nombre": "Mariana Torres",
+      "telefono": "+54 11 6123-4567"
+    },
+    {
+      "nombre": "Diego Ramirez",
+      "telefono": "+54 11 6345-7891"
+    }
+  ]
+}
 
-
-    soup = BeautifulSoup(datos_obtenido.text, "html.parser")
-
-    li_productos = soup.find_all("li", class_="ui-search-layout__item") # li es la etiqueta que contiene los productos
-
+   
+    
+def crear_archivo():
+    with open("datos.json", "w", encoding="utf-8") as file:
+        json.dump(datos_empresa, file)
+    print("Archivo JSON creado con éxito.")     
+        
+    """    
     with open(f"./{nombre_archivo}", "w", encoding="utf-8") as file:
         for li in li_productos:
             titulo = li.find("a", class_="poly-component__title").text.strip()
@@ -23,21 +53,53 @@ def buscar_datos(nombre_archivo="datos.txt"):
             
             
             file.write(titulo + ";" + link + ";" + precio + "\n")
-            
-            
-def agregar_datos():
-    buscar_datos("./nuevos.txt")    
+       """     
+  
+def nuevos_productos():
+    a_buscar = input("Ingrese a buscar: ").replace(" ", "-")
+    URL_BASE = f"https://listado.mercadolibre.com.ar/{a_buscar}"
+    datos_obtenido = requests.get(URL_BASE)
+
+
+    soup = BeautifulSoup(datos_obtenido.text, "html.parser")
+
+    li_productos = soup.find_all("li", class_="ui-search-layout__item") # li es la etiqueta que contiene los productos
+        
+    productos = []
+        
+    for li in li_productos:
+        titulo = li.find("a", class_="poly-component__title").text.strip()
+        precio = li.find("span", class_="andes-money-amount__fraction").text.strip()
+        link = li.find("a", class_="poly-component__title").get("href")
+        
+        productos.append(Producto(titulo, link, precio).__dict__)
     
-    with open("./datos.txt", "a", encoding="utf-8") as archivo_datos, open("./nuevos.txt", "r", encoding="utf-8") as archivos_nuevos:
-        for linea in archivos_nuevos.readlines():
-            archivo_datos.write(linea)
-            
-    os.remove("nuevos.txt")
+    with open("datos.json", "r+",  encoding="utf-8") as file:
+        datos = json.load(file)
+        datos["productos"].append(productos)
+        
+        file.seek(0)  # Mover el cursor al inicio del archivo
+        json.dump(datos, file)   
+         
+def agregar_datos():
+    if os.path.exists("./datos.json"): 
+        nuevos_productos()
+        
+        """
+        with open("./datos.txt", "a", encoding="utf-8") as archivo_datos, open("./nuevos.txt", "r", encoding="utf-8") as archivos_nuevos:
+            for linea in archivos_nuevos.readlines():
+                archivo_datos.write(linea)
+                
+        os.remove("nuevos.txt")
+        """
+        
+    else:
+        crear_archivo()
+        nuevos_productos()
         
          
 def menu():
-    print("1-Crear archivo con datos")
-    print("2-Agregar Productos")
+    print("1-Agregar Productos")
     print("0-Salir")
     opc = int(input("Ingresar opcion: "))
     return opc
@@ -47,14 +109,10 @@ if __name__== "__main__":
     
     while True:
         if opc == 1:
-            buscar_datos()
-        elif opc == 2:
             agregar_datos()
-        elif opc == 3:
+        elif opc == 0:
             break
         else:
             print("Opcion no valida")
         
         opc = menu()
-
-    print("Datos obtenidos y guardados en datos.txt")
